move useWithWalletPassword to the root src
